Switch ced imageonversion to jpeg to webp instead - needs testing

+++ start_files/routes/rentals_scripts.py
@@ -1,4 +1,3 @@
-import io
 from selenium.common.exceptions import WebDriverException, TimeoutException, NoSuchElementException, StaleElementReferenceException
 from start_files.models.mls.rentals_db_section import rentals_sessionLocal, Mls_rentals, replace_old_rentals_db, init_rentals_db
 from selenium.webdriver.chrome.service import Service as ChromeService
@@ -14,6 +13,7 @@ from contextlib import contextmanager
 from dotenv import load_dotenv
 from selenium import webdriver
 from datetime import datetime
+from io import BytesIO
 from PIL import Image
 import requests
 import logging
@@ -85,17 +85,18 @@ def clean_and_rename_directories():
             os.rename(item_path, new_item_path)
 
 def convert_images_to_webp():
-    src_dir = '/var/www/html/fastapi_project/start_files/static/rentals_images/'
-    valid_extensions = ['.jpg', '.jpeg', '.png']
-
-    for filename in os.listdir(src_dir):
-        if any(filename.lower().endswith(ext) for ext in valid_extensions):
-            img_path = os.path.join(src_dir, filename)
-            img = Image.open(img_path)
-            webp_filename = os.path.splitext(filename)[0] + '.webp'
-            img.save(os.path.join(src_dir, webp_filename), 'webp')
-            os.remove(img_path)
-
+   src_dir = '/var/www/html/fastapi_project/start_files/static/rentals_images/'
+   valid_extensions = ['.jpg', '.jpeg', '.png']
+   for listing in os.listdir(src_dir):
+       listing_path = os.path.join(src_dir, listing)
+       if os.path.isdir(listing_path):  
+           for filename in os.listdir(listing_path):
+               if any(filename.lower().endswith(ext) for ext in valid_extensions):
+                   img_path = os.path.join(listing_path, filename)
+                   img = Image.open(img_path)
+                   webp_filename = os.path.splitext(filename)[0] + '.webp'
+                   img.save(os.path.join(listing_path, webp_filename), 'WEBP')
+                   os.remove(img_path)
 
 
 def get_end_time_and_elapsed(start_time):
@@ -267,14 +268,19 @@ def load_page(max_images, min_images, item, timeout, max_retries, delay):
                     for i, img in enumerate(all_imgs[:max_images]):
                         img_url = img.get_attribute('src')
                         if img_url:
-                            response = requests.get(img_url)
-                            image = Image.open(io.BytesIO(response.content))
-                            webp_path = os.path.join(save_dir, f'{i + 1}.webp')
-                            with open(webp_path, 'wb') as file:
-                                file.write(response.content)
-                            image.save(webp_path, 'WEBP')
-                            os.remove(webp_path)  # Delete the original JPG file
-                    logger.info(f"Completed image extraction and conversion for MLS {mls}.")
+                            try:
+                                response = requests.get(img_url)
+                                response.raise_for_status() 
+                                
+                                img = Image.open(BytesIO(response.content))
+                                image_path = os.path.join(save_dir, f'{i + 1}.webp')
+                                img.save(image_path, 'WEBP')
+                                print(f'Saved image as {image_path}')
+                                
+                            except Exception as e:
+                                logger.error(f"Error downloading or saving image {img_url}: {e}")
+
+                    logger.info(f"Completed image extraction for MLS {mls}.")                    
                     
                     db_attempt = 0
                     while db_attempt < max_retries:
@@ -379,7 +385,11 @@ async def start_task_in_loop(concurrency_limit, timeout, max_images, min_images,
 def start_rentals(concurrency_limit, timeout, max_images, min_images, max_price, max_retries, delay):
     start_time = time.time()
     logger.info(f"Start Time: {datetime.fromtimestamp(start_time).strftime('%Y-%m-%d %H:%M:%S')}")
-    prep_directories()
+    try:
+        prep_directories()
+    except Exception:
+        logger.error("No directories found to post-process")
+    
     try:
         loop = asyncio.new_event_loop()
         asyncio.set_event_loop(loop)
@@ -393,8 +403,9 @@ def start_rentals(concurrency_limit, timeout, max_images, min_images, max_price,
 
     replace_old_rentals_db()
     clean_and_rename_directories()
-    logger.info("Database moved to production") 
-    logger.info("Images converted to web format") 
+    logger.info("Database moved to production")     
+    convert_images_to_webp    
+    logger.info("Images converted") 
     elapsed_time = time.time() - start_time
     minutes = int(elapsed_time // 60)
     seconds = elapsed_time % 60
